moved slice utils to util package
