+++ lib/util/slice_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a 2-clause
 // BSD-style license that can be found in the LICENSE file.
 
-package watch
+package util
 
 import (
 	"reflect"
